+++ eg/typewrite.ts
@@ -1,12 +1,12 @@
 import { dev } from '@qqi/dev-log';
-import { typewrite } from 'src/typewrite';
+import { typewrite } from '../index';
 
-dev('测试 typewrite', async it => {
+await dev.skip('测试 typewrite', async it => {
   await it('测试 1', async () => {
     await typewrite('你好');
   });
 
-  await it('测试 慢输出', async () => {
+  it('测试 慢输出', async () => {
     await typewrite('你好，我是 earthnut.dev', 300);
   });