+++ src/run-other-code/parse.ts
@@ -1,7 +1,14 @@
-import { isString } from 'a-type-of-js';
+import {
+  isBoolean,
+  isFalse,
+  isNaN,
+  isNumber,
+  isString,
+  isUndefined,
+} from 'a-type-of-js';
 import { RunOtherCodeOption, RunOtherCodeOptions } from './types';
 import { pathJoin } from '../path';
-import { isNode } from 'a-js-tools';
+import { getRandomInt, isNode } from 'a-js-tools';
 
 /**
  *
@@ -11,8 +18,14 @@ import { isNode } from 'a-js-tools';
 export function parse(options: RunOtherCodeOption): RunOtherCodeOptions & {
   /**  命令组  */
   cmd: string[];
+  /**  waiting  */
+  waiting: {
+    show: boolean;
+    info: string;
+    suffix: number;
+  };
 } {
-  if (!isNode()) {
+  if (isFalse(isNode())) {
     throw new RangeError('当前环境不支持 node 环境');
   }
 
@@ -21,21 +34,62 @@ export function parse(options: RunOtherCodeOption): RunOtherCodeOptions & {
     options = { code: options };
   }
 
+  /**  工作目录  */
   const cwd = pathJoin(process.cwd(), options.cwd || '');
-
+  /**  执行命令  */
   const cmd = options.code
     .replace(/\s{2,}/g, ' ')
     .trim()
     .split(' ');
 
+  const { show, info, suffix } = {
+    show: false,
+    info: '请等待',
+    suffix: getRandomInt(3),
+  };
+  /**  等待  */
+  const waiting = isBoolean(options.waiting)
+    ? {
+        show: options.waiting,
+        info,
+        suffix,
+      }
+    : isString(options.waiting)
+      ? {
+          show: true,
+          info: options.waiting,
+          suffix,
+        }
+      : isUndefined(options.waiting)
+        ? {
+            show: isBoolean(options.hideWaiting) ? !options.hideWaiting : show,
+            info: options.waitingMessage ?? info,
+            suffix,
+          }
+        : isNumber(options.waiting)
+          ? {
+              show: true,
+              info: '请等待',
+              suffix: isNaN(options.waiting)
+                ? suffix
+                : Math.min(Math.max(0, options.waiting), 2),
+            }
+          : {
+              show: true,
+              info,
+              suffix,
+              ...options.waiting,
+            };
+
   /// 混合值，将实参进行整理
   return {
     shell: true,
-    hideWaiting: false,
-    waitingMessage: '',
+    hideWaiting: true,
+    waitingMessage: '请稍等',
     printLog: true,
     ...options,
     cmd,
     cwd,
+    waiting,
   };
 }