+++ src/run-other-code/waiting.ts
@@ -1,4 +1,3 @@
-import { getRandomInt } from 'a-js-tools';
 import { csi } from '@color-pen/static';
 import {
   cursorAfterClear,
@@ -7,41 +6,46 @@ import {
   cursorShow,
 } from '../cursor';
 import { _p } from '../print';
+import { isFalse, isUndefined } from 'a-type-of-js';
+import { RunOtherCodeWaiting } from './types';
 
 /**
  *
  * 等待
  *
  */
-export function waiting(hideWaiting: boolean, waitingMessage: string) {
-  let count = 0;
+export function waitingCn(waiting: RunOtherCodeWaiting) {
+  if (isFalse(waiting.show)) return () => {};
+
   let timeStamp: undefined | NodeJS.Timeout = undefined;
 
   /**  销毁等待信息  */
   function destroyed() {
-    if (timeStamp) {
-      clearInterval(this.count);
+    if (!isUndefined(timeStamp)) {
+      clearInterval(timeStamp);
     }
     cursorMoveLeft(Infinity); /// 移动到最左边
     cursorAfterClear(); /// 🧹光标后的内容，避免出现打印残留
     /// 返回之前将光标展示出来
     cursorShow();
-    process.removeListener('exit', cursorShow);
+    process.removeListener('exit', cursorShow); // 禁止多监听未移除导致程序报错
     cursorShow();
   }
 
-  if (!hideWaiting) {
-    return destroyed;
-  }
+  const pStore = [
+    ['.', '..', '...', '....', '...', '..'],
+    ['···', '⋱', '⋮', '⋰', '···', '⋱', '⋮', '⋰'],
+    ['⤯', '⤰', '⤮', '⤩', '⤪', '⤧', '⤨'],
+    ['🙃', '😂', '😄', '😡', '😏', '🫥', '🥳'],
+    ['👈', '👆', '👉', '👇', '🤘', '🤟', '🫳', '🫴', '👊'],
+  ];
 
   /** 随机出一个待渲染列队 */
-  const pList: string[] = [
-    ['.', '..', '...', '....', '...', '..'],
-    ['···', '⋱', '⋮', '⋰'],
-    ['⤯', '⤰', '⤮', '⤩', '⤪', '⤧', '⤨'],
-  ][getRandomInt(2)];
+  const pList: string[] =
+    pStore[Math.min(pStore.length - 1, Math.max(waiting.suffix, 0))];
   /** 随机出的等待标志符数组的长度 */
   const pLength: number = pList.length;
+  let count = 0;
   /// 隐藏光标
   cursorHide();
   // 放置一个在进程结束时展示光标，即便在测试发现异步操作会阻塞该事件的触发
@@ -51,10 +55,7 @@ export function waiting(hideWaiting: boolean, waitingMessage: string) {
     // 🧹光标后内容
     cursorAfterClear();
     // 打印文本
-    _p(
-      `\n${waitingMessage}${'.'.repeat(++count % pLength)}${csi}20D${csi}1A`,
-      false,
-    );
-  }, 100);
+    _p(`\n${waiting.info}${pList[++count % pLength]}${csi}20D${csi}1A`, false);
+  }, pLength * 16);
   return destroyed;
 }